added a huge (10240x10240) map for testing fixed sfx changed  weapon rof to float improved scene loading (no pixel buffer anymore for the background) fixed collision detection with obstacle grid added character pushing fixed scenerenderer inview method

+++ trunk/src/main/java/it/patamau/ld25/Weapon.java
@@ -16,11 +16,11 @@ public class Weapon {
 	public final String name;
 	private long lastShotTime, shotInterval;
 	public final float recoil, bulletSpeed, bulletSize, bulletDamage, ttl;
-	public final int rof;
+	public final float rof;
 	public final boolean explosive;
 	public final Sfx sfx;
 	
-	private Weapon(final String name, float recoil, float bulletSpeed, float bulletSize, float bulletDamage, int rof, float ttl, boolean explosive, String sfx){
+	private Weapon(final String name, float recoil, float bulletSpeed, float bulletSize, float bulletDamage, float rof, float ttl, boolean explosive, String sfx){
 		this.name = name;
 		this.recoil = recoil;
 		this.bulletSpeed = bulletSpeed;
@@ -32,7 +32,7 @@ public class Weapon {
 		this.sfx = new Sfx(sfx);
 		
 		//internals
-		shotInterval = 1000000000l/rof;
+		shotInterval = (long)(1000000000f/rof);
 		lastShotTime = 0l;
 	}
